+++ aidial_sdk/chat_completion/request.py
@@ -5,9 +5,6 @@ from typing_extensions import assert_never
 
 from aidial_sdk.chat_completion.enums import Status
 from aidial_sdk.deployment.from_request_mixin import FromRequestDeploymentMixin
-from aidial_sdk.deployment.schema_rich_applications_mixin import (
-    SchemaRichApplicationsMixin,
-)
 from aidial_sdk.exceptions import InvalidRequestError
 from aidial_sdk.pydantic_v1 import (
     ConstrainedFloat,
@@ -257,7 +254,6 @@ class ChatCompletionRequest(AzureChatCompletionRequest):
 
 class Request(
     ChatCompletionRequest,
-    FromRequestDeploymentMixin,
-    SchemaRichApplicationsMixin,
+    FromRequestDeploymentMixin
 ):
     pass
