+++ aidial_sdk/deployment/tokenize.py
@@ -2,9 +2,7 @@ from typing import List, Literal, Union
 
 from aidial_sdk.chat_completion.request import ChatCompletionRequest
 from aidial_sdk.deployment.from_request_mixin import FromRequestDeploymentMixin
-from aidial_sdk.deployment.schema_rich_applications_mixin import (
-    SchemaRichApplicationsMixin,
-)
+
 from aidial_sdk.pydantic_v1 import BaseModel
 
 
@@ -21,7 +19,7 @@ class TokenizeInputString(BaseModel):
 TokenizeInput = Union[TokenizeInputRequest, TokenizeInputString]
 
 
-class TokenizeRequest(FromRequestDeploymentMixin, SchemaRichApplicationsMixin):
+class TokenizeRequest(FromRequestDeploymentMixin):
     inputs: List[TokenizeInput]
